fix(context): 修复外部知识扩展边界问题
