+++ frontend/src/__tests__/app/runtime-config-cache.test.ts
@@ -23,6 +23,7 @@ describe('runtime config caching', () => {
         publicApiUrl: '',
         socketDirectUrl: '',
         enableChatContext: true,
+        enableDingTalkContext: false,
         loginMode: 'all',
         oidcLoginText: '',
         enableDisplayQuotas: false,
@@ -72,6 +73,7 @@ describe('runtime config caching', () => {
         publicApiUrl: 'http://1.1.1.1:8000',
         socketDirectUrl: '',
         enableChatContext: true,
+        enableDingTalkContext: false,
         loginMode: 'all',
         oidcLoginText: '',
         enableDisplayQuotas: false,
